requirements.txt
djangorestframework
djangorestframework_simplejwt
python-decouple
psycopg2-binary
Faker
factory-boy